Add files via upload
"This project is a Spring Boot application that calculates and retrieves the distance and duration between two pin codes using the Google Maps API. It includes RESTful APIs for fetching the route information and caching it in a database for quicker future retrievals. The system allows both the saving and fetching of routes, ensuring that frequently used routes are available without external API calls. The project implements caching, database persistence, and external API integration."